Updated OutputSettings inside of Document to be a static inner class.
This addresses the issue with using jsoup and scala 2.8 discussed here:
http://groups.google.com/group/jsoup/browse_thread/thread/3f7ec2fa41dfb87f

This change should be safe to commit and won't make a backwards incompatible
to the public interface of jsoup (you can always reference a static member
via a non-static path) for any existing users.  A recompilation of their code
won't even be necessary.

All tests continue to pass for me after this change.