Make copies of all strings returned, vs returning pointers to substrings of input.
The former method is a little faster, and creates less garbage. However when the input is large,
and apps retain data pulled from the DOM, the app may perceive a memory leak, as even a small string
is actually as large as the original input (although multiple strings are all backed by the one
original input).

So, this implementation is a little less performant, but has a potential for greater safety,
depending on how the library is used.