Modified to save memory when creating many unknown (custom) tags
Previously, new Tag() was created each time a tag with previously used tagName was needed, sometimes creating thousands of Tag instance duplicates. Also, for a custom tag that can "do anything", a choice is needed if it should be formatted as block or not. For this provided a new variant of Tag valueOf(String tagName, boolean formatAsBlock)