Bugfix #951 After clone document
This commit fix the ISSUE https://github.com/jhy/jsoup/issues/951

Problem: Once a document is cloned and try to append an after element the document cloned
doesn't recognize the parent and that result in a nextElementSibling equals to null.

Solution: The Element.clone() method calls Element.doClone() with self as parameter to create a new instance.

Changes:
- Element.clone calls to doClone method with self as arg
- New test case added to verify the generated objects are different and with no null values.

"Omar Bautista" <joxebus@gmail.com>