Start stack search at right spot
Fixes a bug in the in-scope search, which was starting in the middle instead of the bottom when trying to limit the stack depth search. That meant that when generating implied end-tags, the stack wasn't getting shorter, so it tried to generate end tags for days, and ultimately threw a StackOverflowDetect.

(Which is somewhat ironic, because the max depth was set to prevent a different STE... Le Sigh.)

Fixes #966