Add @SafeHtml java validator
This change does not add any new dependencies to jsoup.
If a bean validation implementation is present, then the new
@org.jsoup.constraints.SafeHtml constraints can be used.

This implementation was copied from Hibernate Validator 6.1.5.FINAL
SafeHtml was marked deprecated in Hibernate Validator following
CVE-2019-10219 see https://in.relation.to/2019/11/20/hibernate-validator-610-6018-released/

The root cause of the CVE was a misunderstanding of jsoup by Hibernate
Validator. By making the @SafeHtml constraint part of jsoup, that
concern is eliminated.

With this change, users have a logical source for @SafeHtml,
eliminating the need for each project to build and maintain @SafeHtml
itself, making for more secure and maintainable software.