Performance improvement for clone
The setBaseUri call was redundant now that the base URI is lazy loaded by an element, by walking up the parent stack to find it. But because of that walk, the performance of setBaseUri during clone was ~ O(n^2) or worse.

Also, updated StringUtil to limit padding to 30 wide, otherwise the outerHtml string format made in hasSameValue was blowing the heap for very deep nodes.