Don't try to unconsume after a potential bufferUp when moving from RCData to TagOpen state
Should fix #1542 - I wasn't able to directly test this with the sample fuzz output, but was able to reconstruct the same code execution path and error with a standard unit test.

Also reviewed all 14 other uses of unconsume() and verified that these all are directly after a consume() and there's no bufferUp() potential, so those cannot trigger this underrun.