Improved newline and whitespace normalization
Fixes #1787

In TextNode, if this is a blank, check the next Element if it will indent. If so, can skip. Previously would check if the parent el would indent, which is wrong for inline elements.

Also made empty tags (like <img>) not indent, but inline.

And fixed up how whitespace is normalized at the end of an element, and after the body tag.