Simplified the Test Server startup
Removed the counting latch as Jetty will only start if it's not already running, so the latch was redundant. The stop wasn't being called anyway (which is good - we want the same server to run throughout the test session, vs restarting for each test class).

Also ensure the server binds only to localhost, vs to 0.0.0.0.

Fixes #1822