Remove incoming sibling prior to add in before() and after()
This ensures that the siblingIndex is correct before calling the absolute addChildren(index) to re-position.

Otherwise, if the incoming node was already before the context node (in before()), or after (in after()), the relative order would be incorrect due to being removed and re-indexed prior to re-add. For after(), that could lead to an IndexOutOfBounds exception.

Fixes #1898.