Updated AfterBody and AfterAfterBody to current spec
Don't pop stack to close on </body> or </html>, but leave them on the stack.

Had to deviate from the spec slightly to allow whitespace to be added to the html or doc elements. (The goal of that is so that when pretty-printing is off, the output more closely resembles the input, by tracking newlines after </body> etc)

Fixes #1851