Check the buffer is fully read if looking for absent content
Fixes #1929

In this case we are testing for a missing `</textarea>` - but if the buffer hasn't been fully read, we could never find it.

For the normal case that this code is looking for - a missing `</title>` in brief HTML, a best-effort check (assuming the buffer is complete) is sufficient.