Added a cost-based query planner for CSS selectors
Simpler queries are executed before more complex queries, to improve performance.

Notably, in a query like `div:has(p)`, the element under test is now first evaluated to match the `div` tag, before the deep `:has(p)` scan is executed. Previously it would be the other way around, as the And evaluator always performed a RTL execution.

The initial costs are fairly arbitrary and may be tweaked if required upon further analysis.