Only attempt to correct 8559 -> UTF on response headers
Fixes #2011

This allows header values to be sent as UTF-8 (as encoded by Java HTTP request writer). But that's out of spec and depending on how the upstream server handles that, may or may not work as the user intended. Hence added a note on the docs to suggest encoding the data first.