Disable the fuzz timeout integration tests
These are flaky if the CPU allocation on the GitHub runners is low. And aren't really specified. They are intended to catch regressions in the parse time for (somewhat maliciously crafted) input. Manually execute as performance tests instead for now.

There are other tests that are flaky if the CPU is very throttled - the timeout tests and interupt tests. Hoping that the CPU will get less throttled if we reduce the time of runs overall, and they will come good.