Avoid BufferedInputStream synchronized reads (#2058)
To avoid pinning a virtual thread to a carrier thread, refactored ConstrainedInputStream into ControllableInputStream, which does not extend BufferedInputStream, but wraps it instead.

Had to provide a deprecation path in ConstrainedInputStream as changing the extends causes a binary backwards compatibility issue. Even though folks shouldn't be using Internal package classes, they may be.

Also enabled bodyStream to return a byte array stream after bufferUp(), in case constrains on the stream are still required.

This does not include a test case for pinning. Verified manually. Would be good to find a way to IT test it.