In Cleaner, use the normalized tag name in the tail hit
And ensure that tags added to the safelist are normalized.

In copy safe nodes, head was using normalized but the tail was using nodename. So if the safelist and the input had different cases, the tail hit would not ascend the stack on an otherwise safe tag, leading to mis-nested children.

Also fixed an issue that the cleaned element did not preserve the self-closing style of the input document.

Fixes #2049