Android scents, ignore byteBuffer
While Buffer and ByteBuffer were available since Android 1, incl .flip(), scents was complaining about ControllableInputStream's use of .flip. Best I can see is that because .flip in previous JDKs returned Buffer, and now ByteBuffer, this caused Android Scents to misread the API and flag.

We don't use the return of flip so it's not relevant either way.