[JSOUP-2224] Added wildcards
It is common for applications to use custom attributes, e.g., 'aria-*', and
in fact with HTML 5(?) this is offically recognized with 'data-*' wildcards.
It is often impossible to list all possible custom attributes, yet we want
to be more selective than using ':all'.

This patch adds support for both per-tag and global attribute wildcards.
The wildcard must be recognized by `java.text.Pattern`, e.g., `data-.+`.
The patch will quietly add the '^' and '$' delimiters to the provided
pattern.