Enforce `-release` 8 base compilation target (#2250)
Set -release 8 base compilation target

Restructured the compile phases in the POM so that when executing on Java 9+, the base classes are compiled with -release 8. This ensures that the bootstrap class path uses the Java 8 API, and so the produced jar is binary compat with Java 8.

When built on Java 8, the original source and target (vs release) are used.

Fixes an issue where when running on Java 8 (but built on Java9+),  `java.lang.NoSuchMethodError: java.nio.ByteBuffer.flip()Ljava/nio/ByteBuffer;` would be thrown in ControllableInputStream buffering up, because the return expected `Buffer`, not the binary incompatible `ByteBuffer`.