Reimplement the adoption agency algorithm (#2278)
Fixes #2267

The implementation of the adoption agency algorithm vs the HTML5 spec had drifted over time. This fixes that.

The spec for table fostering when in adoption is not presently implemented, marked as a todo.

Given the complexity of the algo and how it has changed over time, I opted to include it inline so that hopefully later changes are easier to adopt.