Clone tags from source TagSet on demand
This allows the TagSet to vend mutable Tags on demand, without requiring the entire tagset to be cloned on initialization. Otherwise for the HTML tagset, it would create many unused duplicates (all the known tags) on every pass.

Gets the perf to effectively where we were in 1.19.1 with immutable flyweight tags, but with the ability to mutate as required.