Store cached child els in Attributes, not field
We keep a lazy initialized, cached copy of an Element's child Element list, to save recreating it every time they are requested.

That was stored in a nullable field. This change moves it into the userdata section of the attributes instead.

Also, rather than tracking changes to the backing node list via a ChangeNotifyingArrayList, I'm using the ArrayList's modcount instead directly. So we can deprecate the ChangeNotifyingArrayList.

Ultimately this trims the shallow size of an Element from 40 to 32 bytes, and the NodeList from 32 to 24 (20% and 25% decreases respectively)