Consume escaped subqueries correctly
Now that escaped sequences can have spaces, that needs to be handled in consumeSubQuery correctly.

The unescape -> escape step is a bit redundant. If perf critical we could have a method that consumes escape sequences without actually unescaping them. Or, we could refactor consumeSubQuery to parse into Evaluators directly.

But the rest is already pretty stringy, and users can keep parsed Evaluators for reuse, so am deeming OK for now.

Fixes issue noted in #2305