Retain same Parser in Document & Element clone
When cloning an Element or a Document, the same underlying Parser should be retained.

We were creating new Parsers with HTML treebuilders on each Element clone. Which is both heavyweight and incorrect.

The Parser is now threadsafe so if the cloned element was being used (appending HTML) across threads, this will still be safe. (Albeit not concurrent; users can set a new Parser via ownerDocument().parser().newInstance() for a copy if required.)