Make Parser instances threadsafe (#2314)
* Make Parser instances threadsafe

I've seen examples of people trying to use a Parser in a concurrent manner, and this may succeed in low-concurrency modes, but will fail when under pressure. Adding an explicit lock to ensure safety.

* Retain same Parser in Document & Element clone (#2315)

When cloning an Element or a Document, the same underlying Parser should be retained.

We were creating new Parsers with HTML treebuilders on each Element clone. Which is both heavyweight and incorrect.

The Parser is now threadsafe so if the cloned element was being used (appending HTML) across threads, this will still be safe. (Albeit not concurrent; users can set a new Parser via ownerDocument().parser().newInstance() for a copy if required.)