Added IOException throwing `readFully` and `readBody` (#2327)
For historical reasons (specifically that a response's body used to be fully buffered during `execute`), `Connection#body()` does not throw a checked IOException.

This change adds a new `Connection#readBody()` which does throw a checked exception, and a partner `Connection#readFully()` method to do the internal buffering.

Clarified relevant documentation.