Performance optimizations for some DOM methods
The primary optimization is to invalidate sibling indexes on modification, vs re-calculating them immediately. When called in a loop of `element.child(0).remove()`, the previous implementation was effectively O(n²), and is now just O(n).

Also includes optimizations for `Document.createElement()`, setting attributes, and when parsing a body fragment with many direct children.